+++ src/main/java/jpa/repository/FavoriteRepository.java
@@ -1,4 +1,4 @@
-package jpa.dao;
+package jpa.repository;
 
 import jpa.domain.Favorite;
 import org.springframework.data.jpa.repository.JpaRepository;
