+++ src/main/java/jpa/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package jpa.dao;
+package jpa.repository;
 
 import jpa.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
