+++ src/test/java/jpa/repository/LineRepositoryTest.java
@@ -1,9 +1,6 @@
-package jpa.dao;
+package jpa.repository;
 
 import jpa.domain.Line;
-import jpa.domain.LineStation;
-import jpa.domain.Station;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
